Fix trim whitespaces in add-to-one-line-help
Trimming whitespaces via the awk command worked but also removes
multiple whitespaces between columns. Use a sed command instead.
I also tested the bash variable substitution/expansion commands, but
they appear not to work in scripts, but do work on the repl.
The commands that I tried are the following:

  my_var='  Something untrimmed   '
  echo "${my_var##*( )}"  # trim all leading spaces; apparently only works on the repl
  echo "${my_var%%*( )}!"     # trim all trailing spaces; may only work on the repl